Dedupe recreate-class resets and extract executor routing module
orchestrator.ts has a 5500-line guardrail and master was already at 5493,
so the recreateDownstream addition pushed it over. Shrink the file by:
- sharing one RECREATE_RESET_CHANGES payload + applyRecreateReset tail
  across recreateTask / recreateDownstream / recreateWorkflow
- moving the self-contained executor routing rules (pure functions, no
  class state) to executor-routing.ts, re-exported from orchestrator.ts
  so the public API is unchanged
